+++ src/client/components/tl_display/tl_display.js
@@ -17,14 +17,18 @@ Template.tl_display.onRendered( function(){
     const count = Template.currentData().msg.count;
     //console.debug( 'count', count );
 
-    if( count === 0 ){
-        self.$( '.tl-display .alert[data-tl-count="'+count+'"]' ).css({ top: CSS_TOP });
-    
-    } else {
-        let prevTop = parseInt( $( '.tl-display .alert[data-tl-count="'+( count-1 )+'"]' ).css( 'top' ));
-        let prevHeight = parseInt( $( '.tl-display .alert[data-tl-count="'+( count-1 )+'"]' ).css( 'height' ));
-        self.$( '.tl-display .alert[data-tl-count="'+count+'"]' ).css({ top: (prevTop+prevHeight)+CSS_MARGIN+'px' });
+    const start = tlTolert.conf.stackFromTop ? 'top' : 'bottom';
+    let args = {};
+    let distance = CSS_TOP;
+
+    if( count > 0 ){
+        const prevTop = parseInt( $( '.tl-display .alert[data-tl-count="'+( count-1 )+'"]' ).css( start ));
+        const prevHeight = parseInt( $( '.tl-display .alert[data-tl-count="'+( count-1 )+'"]' ).css( 'height' ));
+        distance = (prevTop+prevHeight) + CSS_MARGIN + 'px';
     }
+
+    args[start] = distance;
+    self.$( '.tl-display .alert[data-tl-count="'+count+'"]' ).css( args );
 
     // remove the message from our local collection after the configured time
     if( tlTolert.conf.timeout > 0 ){